chore: restructuring intractions
